SXC Shader Permutation Blobs (#20)
* move into external

* write and read permutation blobs

* add metadata file

* move into one namespace

* fix metadata file type

* Update README.md

* add missing tbb redist files

+++ main.cpp
@@ -150,7 +150,7 @@ int main(int argc, char* argv[])
 
         // Print duration in seconds with milliseconds
         auto ms = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count();
-        double seconds = ms / 1000.0;
+        double seconds = static_cast<double>(ms) / 1000.0;
 
         printf("========== Build completed and took %.3f seconds ==========\n", seconds);
